+++ lifters-log-rewrite/src/app/routes/Social.tsx
@@ -2,13 +2,16 @@ import Chat from "../../features/chat/components/Chat";
 import FriendActivity from "../../features/friends/components/FriendActivity";
 import "../../features/chat/chat.css";
 import "../../features/friends/friends.css";
+import ChatReplyContextProvider from "../../features/chat/contexts/ChatReplyContext";
 
 function Social() {
     return (
         <>
             <div className="mainContentPane">
-                <Chat />
-                <FriendActivity />
+                <ChatReplyContextProvider>
+                    <Chat />
+                    <FriendActivity />
+                </ChatReplyContextProvider>
             </div>
         </>
     )