+++ lifters-log-rewrite/src/features/chat/components/ChatMessage.tsx
@@ -3,8 +3,9 @@ import TimeSubtext from "../../../components/TimeSubtext";
 import useHoverTouch from "../../../hooks/useHoverTouch";
 import UserData from "../../../types/UserData";
 import { PLACEHOLDERUSERDATA } from "../../../utils/constants";
-import { motion, PanInfo, useMotionValue } from "framer-motion";
+import { motion, useMotionValue } from "framer-motion";
 import { useEffect } from "react";
+import { useReplying } from "../contexts/ChatReplyContext";
 
 type ChatMessageProps = {
     cid: number;
@@ -12,14 +13,13 @@ type ChatMessageProps = {
     author: string;
     time: string;
     date: string;
-    repliesTo?: {author: string, message: string};
-    setReplyingMessageId: (rid: number) => void;
-    setReplyingMessageText: (text: string) => void;
+    repliesTo?: {author: string, message: string, type: "note" | "lift" | "cardio" | "label" | "message" | undefined};
 }
 
-function ChatMessage({ cid, msg, author, time, date, repliesTo, setReplyingMessageId, setReplyingMessageText }: ChatMessageProps) {
+function ChatMessage({ cid, msg, author, time, date, repliesTo }: ChatMessageProps) {
 
     const { isHovering, hoverHandlers } = useHoverTouch();
+    const { setReplyingId, setReplyingText, setReplyType, setOriginalUser } = useReplying();
     const authUser = useAuthUser<UserData>() || PLACEHOLDERUSERDATA;
     const x = useMotionValue(0);
 
@@ -29,6 +29,19 @@ function ChatMessage({ cid, msg, author, time, date, repliesTo, setReplyingMessa
     const messageDate = new Date(Date.parse(date + "T00:00:00"))
     // Date.parse() parses UTC strings then converts them to your timezone which can mess up dates
     // Adding a time makes it parse it in local time
+
+    let replyClass = "";
+    if (repliesTo) {
+        if (repliesTo.type === "note") {
+            replyClass = "noteActivity"
+        } else if (repliesTo.type === "lift") {
+            replyClass = "logActivity"
+        } else if (repliesTo.type === "cardio") {
+            replyClass = "cardioActivity"
+        } else if (repliesTo.type === "label") {
+            replyClass = "labelActivity"
+        }
+    }
 
     useEffect(() => {
         const unsubscribe = x.on("change", (latest) => {
@@ -42,8 +55,10 @@ function ChatMessage({ cid, msg, author, time, date, repliesTo, setReplyingMessa
     }, [x])
 
     function setActiveReply() {
-        setReplyingMessageId(cid)
-        setReplyingMessageText(`${author}: ${msg}`)
+        setReplyingId(cid)
+        setReplyingText(`${author}: ${msg}`)
+        setReplyType("message")
+        setOriginalUser(author)
     }
 
     return (
@@ -60,10 +75,16 @@ function ChatMessage({ cid, msg, author, time, date, repliesTo, setReplyingMessa
             <span style={{fontWeight: "bold"}}>{author}</span>
             {
                 repliesTo &&
-                    <div className={`repliedMessage ${isUsersMessage ? "userRepliedMessage" : ""}`}>
+                    <div className={`repliedMessage ${isUsersMessage ? "userRepliedMessage" : ""} ${replyClass != "" && replyClass}`}>
                         <span style={{fontWeight: "bold"}}>{repliesTo.author}</span>
                         <br />
-                        {repliesTo.message}
+                        <div style={{
+                            whiteSpace: "pre-wrap", 
+                            paddingLeft: `${repliesTo.type != "message" ? "1rem" : "0"}`,
+                            fontStyle:  `${repliesTo.type != "message" ? "italic" : "normal"}`
+                            }}>
+                            {repliesTo.message}
+                        </div>
                     </div>
             }
             <div